+++ components/custom/StaticChip/StaticChip.svelte
