+++ components/custom/StaticChip/index.js
@@ -1,0 +1,3 @@
+import StaticChip from './StaticChip.svelte'
+
+export default StaticChip
