+++ components/custom/index.js
@@ -1,0 +1,5 @@
+import Badge from './Badge'
+import Form from './Form'
+import StaticChip from './StaticChip'
+
+export { Badge, Form, StaticChip }
