+++ components/mdc/CustomCard/index.js
@@ -1,0 +1,3 @@
+import CustomCard from './CustomCard.svelte'
+
+export default CustomCard
