+++ components/mdc/Tour/index.js
@@ -1,0 +1,3 @@
+import Tour from './Tour.svelte'
+
+export default Tour
