+++ stories/Badge.stories.svelte
@@ -1,6 +1,7 @@
 <script>
 import { Meta, Template, Story } from '@storybook/addon-svelte-csf'
-import { Badge } from '../'
+import { Badge } from '../components/custom'
+
 import { copyAndModifyArgs } from './helpers.js'
 
 const args = {
