+++ stories/CustomCard.stories.svelte
@@ -1,6 +1,6 @@
 <script>
 import { Meta, Template, Story } from '@storybook/addon-svelte-csf'
-import { CustomCard } from '../'
+import { CustomCard } from '../components/mdc'
 import { copyAndModifyArgs } from './helpers.js'
 
 const args = {
